feat: extraction-AUC replication gate + md.md steering
Gate 1 of the Anthropic Persona Vectors replication: leave-one-out AUC on
contrastive prompt pairs. A vector that generalizes beyond its own prompt
set will score ~1.0; a memorizer collapses to ~0.5.

- vectors/eval.py: extraction_auc_loo, extraction_auc_all, rank-based AUC
- scripts/run_extraction_auc.py: CLI driver, table + JSON output
- md.md: self-modifying steering file (from development/md.md/) seeded
  with current goal, next action, and smoke-test results
- torch_dtype -> dtype throughout, for transformers 5.x compatibility

Smoke tests (CPU, 2026-04-24):
- gpt2: 1/9 traits pass AUC >= 0.85 (curiosity only). Expected; not instruct.
- Qwen-2.5-0.5B-Instruct: 4/9 pass — joy, curiosity, surprise, caregiver.
  Character traits (honesty, sycophancy, hallucination) need a bigger model.
Next decisive test: Qwen-2.5-7B-Instruct on CARC with the layer sweep
already scaffolded in experiments/00_replication/run.sh.

+++ src/intrinsic_agents/agents/llm_agent.py
@@ -31,7 +31,7 @@ class LLMAgent:
         tok = AutoTokenizer.from_pretrained(model_name)
         if tok.pad_token is None:
             tok.pad_token = tok.eos_token
-        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype=dtype).to(device)
+        model = AutoModelForCausalLM.from_pretrained(model_name, dtype=dtype).to(device)
         probe = ActivationProbe.from_cache_dir(cache_dir, model_name, layer=layer)
         probe.attach(model)
         return cls(name=name, model=model, tokenizer=tok, probe=probe)
